backend/requirements.txt
crewai
crewai_tools
langchain_google_genai
fastapi
uvicorn
python-dotenv
pydantic
